requirements.txt
typer
ipython
feedparser
atproto
jsonpickle
python-handlebars
